+++ Source/NativeApi/NativeApi/Api/Application/DragEventData.cs
@@ -7,8 +7,8 @@ namespace NativeApi.Api
     {
         public IntPtr data;
 
-        public double mouseClientLocationX;
-        public double mouseClientLocationY;
+        public float mouseClientLocationX;
+        public float mouseClientLocationY;
 
         public DragDropEffects effect;
     }
